% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multe.R
\name{multe}
\alias{multe}
\title{Multiple Treatment Effects Regression}
\usage{
multe(
  formula,
  data,
  subset,
  na.action,
  vce = "robust",
  base_val,
  decomp = FALSE,
  minmax = FALSE,
  alpha = 0.05,
  save_lambda_as,
  save_tau_as,
  print = TRUE
)
}
\arguments{
\item{formula}{object of class \code{"formula"} (or one that can be coerced
to that class) of the form \code{outcome ~ treatment | control}}

\item{data}{optional data frame, list or environment (or object coercible by
\code{as.data.frame} to a data frame) containing the outcome and running
variables in the model. If not found in \code{data}, the variables are
taken from \code{environment(formula)}, typically the environment from
which the function is called.}

\item{subset}{optional vector specifying a subset of observations to be used
in the fitting process.}

\item{na.action}{function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{options} (usually \code{na.omit}). Another possible value is
\code{na.fail}}

\item{vce}{Specifies the type of standard errors to report, it can be a
string equal to
\describe{
   \item{"robust"}{Heteroskedasticity-robust standard errors (default).}

   \item{"oracle"}{Heteroskedasticity-robust standard errors, treating
   propensity scores as known.}

}}

\item{base_val}{Specifies the baseline value of the treatment variable,
the first unique value of the treatment variable is used as default.}

\item{decomp}{Compute contamination bias decomposition.}

\item{minmax}{Save saturated group-specific treatment effects as tau
and/or implicit ATE regression weights as lambda; computes bias
decomposition internally.}

\item{alpha}{Determines the confidence level, \code{1-alpha} for
constructing/optimizing confidence intervals.}

\item{save_lambda_as}{Saves the set of implicit ATE regression weights,
with the input as the prefix of variable names. The saved data frame
can be merged back to the data set via index.}

\item{save_tau_as}{Saves the saturated group-specific treatment effects,
with the input as the prefix of variable names. The saved data frame
can be merged back to the data set via index.}

\item{print}{print a summary of the results.}
}
\value{
Returns a summary of the results, containing
   \describe{
     \item{\code{estimation}}{For treatment effect estimation results, where
        \describe{
         \item{\code{est}}{Point estimates of the treatment effect, via 3
              methods: ATE (\code{est[,1]}), one-at-a-time (\code{est[,2]}),
             and common weights (\code{est[,3]}).}

         \item{\code{se_po}}{Heteroskedasticity-robust standard errors.}

         \item{\code{se_or}}{Oracle standard errors: heteroskedasticity-robust,
             treating propensity scores as known.}

         \item{\code{po_vcov}}{Variance-covariance matrix, heterogeneity-robust.}

         \item{\code{or_vcov}}{Variance-covariance matrix, oracle.}
}}
     \item{\code{decomposition}}{For contamination bias decomposition
       results, where
       \describe{
         \item{\code{est}}{Point estimates of the contamination bias decomposition,
             including: coefficients (\code{est[,1]}), own effect
             (\code{est[,2]}), contamination bias (\code{est[,3]}), and its
             maximum (\code{est[,4]}) and minimum (\code{est[,5]}).}

         \item{\code{se}}{Standard errors.}

         \item{\code{lambda_saved}}{Saved lambdas, the set of implicit ATE
             regression weights. Only appear when \code{save_lambda_as}
             is specified.}

         \item{\code{tauhat_saved}}{Saved lambdas, the saturated
             group-specific treatment effects. Only appear when
             \code{save_tau_as} is specified.}
}}
}
   And in both \code{estimation} and \code{decomposition},
    \describe{
  \item{\code{n_obs}}{Number of effect observations used for analysis (with
             weak overlapping strata dropped.}

  \item{\code{n_trt}}{Number of treatment arms.}

  \item{\code{Tlevels}}{The label vector treatment arms.}

}
   If \code{decomp} is not specified, only returns \code{estimation} as a
   list.
}
\description{
Calculate estimators adjusted for the contamination bias in multiple
treatment setting with heterogeneous treatment effects.
}
\section{Note}{

\code{subset} is evaluated in the same way as variables in \code{formula},
that is first in \code{data} and then in the environment of \code{formula}.
}

\examples{

# Project STAR dataset
multe(score ~ treatment | school, data = star)
multe(score ~ treatment | school, data = star, decomp = TRUE, minmax = TRUE)
multe(score ~ treatment | school, data = star, decomp = TRUE, save_lambda_as = "lambda")

}
\references{
{

\cite{Goldsmith-Pinkham, Paul, Peter Hull, and Michal Kolesár. Contamination
bias in linear regressions. No. w30108. National Bureau of Economic
Research, 2022.
\doi{10.3386/w30108}}

}
}
