% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multeR-package.R
\docType{package}
\name{multeR-package}
\alias{multeR}
\alias{multeR-package}
\title{multeR: Multiple Treatment Effects Regression}
\description{
Calculate estimators adjusted for the contamination bias in multiple treatment setting with heterogeneous treatment effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SaiChrisZHANG/multe-R}
  \item Report bugs at \url{https://github.com/SaiChrisZHANG/multe-R/issues}
}

}
\author{
\strong{Maintainer}: Sai Zhang \email{saizhang.econ@gmail.com} [translator]

Authors:
\itemize{
  \item Paul Goldsmith-Pinkham \email{paul.goldsmith-pinkham@yale.edu}
  \item Peter Hull \email{peter_hull@brown.edu}
  \item Michal Kolesár \email{kolesarmi@googlemail.com}
}

}
\keyword{internal}
